requirements.txt
fastapi
langchain
langchain-openai
uvicorn
pydantic
python-multipart
openai
python-dotenv